Creating routes for beers and users
